requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
dacite==1.8.1
dataclasses-json==0.6.7
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-google-community==1.0.7
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.98
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.40.3
orjson==3.10.7
packaging==24.1
pkg_resources==0.0.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
youtube-search==2.1.2
